moved all the functions related to dump log to a separate file
exposed some of the threadpools flags for this purpose

also exposed turnon, turnoff and isset functions, and get the threadpool's name

added unsafe get num threads alive function
resume is counting backwards now

moved functions that acts on the threadpools from thpool_api to the new file
